refactor jsapi - logbypass
