+++ binding.gyp
@@ -49,6 +49,7 @@
                 "src/jsapi/export_logger.cc",
                 "src/jsapi/export_hooks.cc",
                 "src/jsapi/export_http.cc",
+                "src/jsapi/export_thread_logbypass.cc",
             ],
             "include_dirs": [
                 'src',
