+++ src/jsapi/include/export_thread_logbypass.h
@@ -1,0 +1,10 @@
+#ifndef XPROFILER_SRC_JSAPI_THREAD_LOGBYPASS_H
+#define XPROFILER_SRC_JSAPI_THREAD_LOGBYPASS_H
+
+#include "nan.h"
+
+namespace xprofiler {
+void RunLogBypass(const Nan::FunctionCallbackInfo<v8::Value>& info);
+}
+
+#endif /* XPROFILER_SRC_JSAPI_THREAD_LOGBYPASS_H */
