+++ src/xprofiler.cc
@@ -4,8 +4,8 @@
 #include "jsapi/include/export_hooks.h"
 #include "jsapi/include/export_http.h"
 #include "jsapi/include/export_logger.h"
+#include "jsapi/include/export_thread_logbypass.h"
 #include "library/common.h"
-#include "logbypass/log.h"
 #include "nan.h"
 #include "platform/platform.h"
 #include "process_data.h"
